+++ src/Jobs/SendAllDeviceJob.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Tocaanco\FcmFirebase\Jobs;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Tocaanco\FcmFirebase\Facades\FcmFirebase;
+use Illuminate\Contracts\Queue\ShouldBeUnique;
+
+class SendAllDeviceJob implements ShouldQueue
+{
+    use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
+    public $data = [];
+
+    /**
+     * Create a new job instance.
+     *
+     * @return void
+     */
+    public function __construct($data)
+    {
+        //
+        $this->data = $data;
+    }
+
+    /**
+     * Execute the job.
+     *
+     * @return void
+     */
+    public function handle()
+    {
+        FcmFirebase::sendToAllDevices($this->data);
+    }
+}