+++ src/Traits/FcmDeviceTrait.php
@@ -4,7 +4,7 @@ namespace Tocaanco\FcmFirebase\Traits;
 
 trait FcmDeviceTrait
 {
-    public function devices()
+    public function deviceTokens()
     {
         return $this->hasMany(config("fcm-firebase.model"), config("fcm-firebase.user_colum"));
     }