+++ changelog.md
@@ -6,7 +6,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## Version 1.0
 
 ### Added
-- Everything
+- Add model migration for device token table 
+- Add service to handle the logic for fcm
+- Add job to send to all devices token in our database
+- Add channel to notification
 
 ### Changed
 - Nothing